Add files via upload
In this file there's a html code of which started the whole website i have created right now which was like the bones to the human body, CSS is also there in the website which acted like the skin to my website. My website in particular is where women and girls can feel safe enough to talk about their problems and the anonymity of it makes it better because none of the women know each other and having that as the case makes you feel better knowing that the person doesn't know but you managed to talk to them without them judging you "there's no better listener than a stranger" that's the code that inspired the whole website.